# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Naval Propulsion Condition Monitoring API",
    "description": "Backend API for decay prediction and maintenance decision support.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/predict/decay": {
      "post": {
        "summary": "Predict Decay",
        "operationId": "predict_decay_predict_decay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecayPrediction"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/maintenance/recommend": {
      "post": {
        "summary": "Recommend Maintenance",
        "operationId": "recommend_maintenance_maintenance_recommend_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DecayPrediction": {
        "properties": {
          "compressor_decay": {
            "type": "number",
            "title": "Compressor Decay"
          },
          "turbine_decay": {
            "type": "number",
            "title": "Turbine Decay"
          },
          "severity": {
            "type": "string",
            "enum": [
              "critical",
              "warning",
              "caution",
              "healthy"
            ],
            "title": "Severity"
          },
          "operating_condition": {
            "type": "string",
            "title": "Operating Condition"
          },
          "prediction_confidence": {
            "type": "number",
            "title": "Prediction Confidence",
            "description": "Reference model quality from report (R^2 approx.)"
          }
        },
        "type": "object",
        "required": [
          "compressor_decay",
          "turbine_decay",
          "severity",
          "operating_condition",
          "prediction_confidence"
        ],
        "title": "DecayPrediction"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MaintenanceRequest": {
        "properties": {
          "compressor_decay": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.9,
            "title": "Compressor Decay"
          },
          "turbine_decay": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.9,
            "title": "Turbine Decay"
          }
        },
        "type": "object",
        "required": [
          "compressor_decay",
          "turbine_decay"
        ],
        "title": "MaintenanceRequest"
      },
      "MaintenanceResponse": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "immediate",
              "scheduled",
              "monitor"
            ],
            "title": "Action"
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Priority"
          },
          "components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Components"
          },
          "maintenance_window": {
            "type": "string",
            "title": "Maintenance Window"
          }
        },
        "type": "object",
        "required": [
          "action",
          "priority",
          "components",
          "maintenance_window"
        ],
        "title": "MaintenanceResponse"
      },
...
      "PredictRequest": {
        "properties": {
          "ship_speed": {
            "type": "integer",
            "maximum": 27.0,
            "minimum": 3.0,
            "title": "Ship Speed",
            "description": "Ship speed in knots"
          },
          "lever_pos": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Lever Pos",
            "description": "Lever position"
          }
        },
        "type": "object",
        "required": [
          "ship_speed",
          "lever_pos"
        ],
        "title": "PredictRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}